Moving the local auditing module
To a new name to not conflict with the aliasing of `akd_core::proto` which contains the actual protobuf specifications necessary for type conversions.
