+++ akd/src/local_auditing.rs
